refactor: add use unseen count hook
