+++ packages/notification-center/src/components/notification-center/components/FeedsTabs.tsx
@@ -1,12 +1,11 @@
 import React, { useContext, useEffect, useState } from 'react';
-import { Center, Tab } from '@mantine/core';
+import { Tab } from '@mantine/core';
 import styled from 'styled-components';
 import { NotificationsListTab } from './NotificationsListTab';
-import { INotificationCenterContext } from '../../../index';
+import { INotificationCenterContext, useUnseenCount } from '../../../index';
 import { UnseenBadge } from './UnseenBadge';
-import { UnseenCountContext } from '../../../store/unseen-count.context';
 import { Tabs } from './layout/tabs/Tabs';
-import { NotificationCenterContext } from '../../../store/notification-center.context';
+import { NotificationCenterContext } from '../../../store';
 import { useApi, useNovuContext } from '../../../hooks';
 
 export function FeedsTabs() {
@@ -46,7 +45,7 @@ const TabLabelWrapper = styled.div`
 function UnseenBadgeContainer({ storeId }: { storeId: string }) {
   const { api } = useApi();
   const { stores } = useNovuContext();
-  const { unseenCount: generalUnseenCount } = useContext(UnseenCountContext);
+  const { unseenCount: generalUnseenCount } = useUnseenCount();
 
   const [unseenCount, setUnseenCount] = useState<number>();
