+++ packages/notification-center/src/components/popover-notification-center/PopoverNotificationCenter.tsx
@@ -1,11 +1,10 @@
-import React, { useContext } from 'react';
+import React from 'react';
 import { IMessage, IMessageAction, ButtonTypeEnum } from '@novu/shared';
 import { NotificationCenter } from '../notification-center';
 import { INotificationBellProps } from '../notification-bell';
 import { Popover } from './components/Popover';
-import { UnseenCountContext } from '../../store/unseen-count.context';
-import { INovuThemePopoverProvider } from '../../store/novu-theme-provider.context';
-import { useDefaultTheme } from '../../hooks';
+import { INovuThemePopoverProvider } from '../../store';
+import { useDefaultTheme, useUnseenCount } from '../../hooks';
 import { ColorScheme, ListItem, ITab } from '../../index';
 
 interface IPopoverNotificationCenterProps {
@@ -26,7 +25,7 @@ interface IPopoverNotificationCenterProps {
 
 export function PopoverNotificationCenter({ children, ...props }: IPopoverNotificationCenterProps) {
   const { theme } = useDefaultTheme({ colorScheme: props.colorScheme, theme: props.theme });
-  const { setUnseenCount, unseenCount } = useContext(UnseenCountContext);
+  const { setUnseenCount, unseenCount } = useUnseenCount();
 
   function handlerOnUnseenCount(count: number) {
     if (isNaN(count)) return;
