+++ packages/notification-center/src/hooks/use-novu-theme-provider.hook.ts
@@ -1,6 +1,5 @@
 import { useContext } from 'react';
-import { INovuTheme, ThemeContext } from '../store/novu-theme.context';
-import { ICommonTheme } from '../store/novu-theme-provider.context';
+import { INovuTheme, ThemeContext, ICommonTheme } from '../store';
 import { ColorScheme } from '../index';
 
 export function useNovuThemeProvider(): {
