+++ packages/notification-center/src/hooks/use-unseen-count.hook.ts
@@ -1,24 +1,8 @@
-import { useEffect, useState } from 'react';
-import { useApi } from './use-api.hook';
-import { useAuth } from './use-auth.hook';
+import { useContext } from 'react';
+import { UnseenCountContext } from '../store';
 
-export function useUnseenController() {
-  const [unseenCount, setUnseenCount] = useState<number>(0);
-  const { api } = useApi();
-  const { token } = useAuth();
+export function useUnseenCount() {
+  const { unseenCount, setUnseenCount } = useContext(UnseenCountContext);
 
-  useEffect(() => {
-    if (!token || !api?.isAuthenticated) return;
-
-    (async () => {
-      const { count } = await api.getUnseenCount();
-
-      setUnseenCount(count);
-    })();
-  }, [token, api?.isAuthenticated]);
-
-  return {
-    unseenCount,
-    setUnseenCount,
-  };
+  return { unseenCount, setUnseenCount };
 }
